refactor(cmake): update library dependencies for system model
