+++ src/expressions/CMakeLists.txt
@@ -82,6 +82,7 @@ add_library(Antares::expressions ALIAS expressions)
 target_link_libraries(expressions
         PUBLIC
         Boost::headers
+        Antares::system-model-base
         Antares::antares-study-system-model
         Antares::logs
         Antares::vector-utils
