+++ src/expressions/include/antares/expressions/nodes/ParameterNode.h
@@ -3,7 +3,7 @@
 #include <string>
 
 #include <antares/expressions/nodes/Leaf.h>
-#include <antares/study/system-model/variabilityType.h>
+#include <antares/study/system-model-base/variabilityType.h>
 
 namespace Antares::Expressions::Nodes
 {
