+++ src/expressions/include/antares/expressions/nodes/VariableNode.h
@@ -3,7 +3,7 @@
 #include <string>
 
 #include <antares/expressions/nodes/Leaf.h>
-#include <antares/study/system-model/variabilityType.h>
+#include <antares/study/system-model-base/variabilityType.h>
 
 namespace Antares::Optimisation
 {
