+++ src/study/CMakeLists.txt
@@ -1,3 +1,4 @@
 OMESSAGE("Antares Study")
 
+add_subdirectory(system-model-base)
 add_subdirectory(system-model)
